feat: update authentication flow to use Google sign-in and remove legacy sign-in pages
