+++ .env.local.example
@@ -2,3 +2,5 @@ AUTH_SECRET="your_auth_secret_here"
 NEXTAUTH_URL="http://localhost:3000"
 POSTGRES_PRISMA_URL=postgresql://postgres:postgres@localhost:7777/postgres
 POSTGRES_URL_NON_POOLING=postgresql://postgres:postgres@localhost:7777/postgres
+GOOGLE_CLIENT_ID="your_google_client_id_here"
+GOOGLE_CLIENT_SECRET="your_google_client_secret_here"
