--- app/login/signup/page.tsx
@@ -1,92 +1,0 @@
-"use client";
-
-import { signIn, useSession } from "next-auth/react";
-import { useRouter } from "next/navigation";
-import { useEffect } from "react";
-
-export default function SignUpPage() {
-  const { data: session, status } = useSession();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (status === "authenticated") {
-      router.replace("/");
-    }
-  }, [status, router]);
-
-  if (status === "loading") {
-    return (
-      <div className="flex items-center justify-center min-h-screen bg-gray-50">
-        <div className="h-12 w-12 animate-spin rounded-full border-4 border-blue-500 border-t-transparent"></div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex items-center justify-center min-h-screen bg-gray-50">
-      <div className="p-8 bg-white rounded-lg shadow-md w-96">
-        <h1 className="text-2xl font-bold mb-6 text-center">Sign Up</h1>
-        <p className="text-gray-600 mb-6 text-center">
-          Create your account to start practicing Japanese characters. 
-          We'll automatically set up your account using Google.
-        </p>
-        
-        <button
-          onClick={() => signIn("google")}
-          className="w-full flex items-center justify-center gap-3 px-6 py-3 bg-white border border-gray-300 rounded-md hover:bg-gray-50 transition-colors font-medium text-gray-700 mb-4"
-        >
-          <svg
-            className="w-5 h-5"
-            viewBox="0 0 24 24"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <path
-              fill="#4285F4"
-              d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
-            />
-            <path
-              fill="#34A853"
-              d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
-            />
-            <path
-              fill="#FBBC05"
-              d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
-            />
-            <path
-              fill="#EA4335"
-              d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
-            />
-          </svg>
-          Sign Up with Google
-        </button>
-
-        <div className="text-center">
-          <p className="text-sm text-gray-600 mb-4">
-            Already have an account?{" "}
-            <button
-              onClick={() => router.push("/login")}
-              className="text-blue-600 hover:text-blue-800 font-medium"
-            >
-              Sign In
-            </button>
-          </p>
-          
-          <button
-            type="button"
-            className="w-full px-6 py-3 bg-gray-200 text-black rounded-md hover:bg-gray-300 transition-colors font-semibold"
-            onClick={() => router.push("/login")}
-          >
-            Back
-          </button>
-        </div>
-
-        <div className="mt-6 text-xs text-gray-500 text-center">
-          <p>
-            By signing up, you agree to our terms of service and privacy policy.
-            Your Google account information will be used to create your profile.
-          </p>
-        </div>
-      </div>
-    </div>
-  );
-}
